Added RepositoriesDB to Notification, Portfolio, updated a bit the other repositories and added the tests

+++ src/main/java/org/example/model/ValueStock.java
@@ -12,7 +12,7 @@ public class ValueStock extends Stock{
         return dividend_rate;
     }
 
-    public void setDividend_rate(int dividend_rate) {
+    public void setDividend_rate(double dividend_rate) {
         this.dividend_rate = dividend_rate;
     }
 }
